+++ openvibe-plugins/simple-visualisation/trunc/src/ovpCFrequencySpectrumDisplay/ovpCFrequencySpectrumDisplayView.cpp
@@ -189,7 +189,7 @@ namespace OpenViBEPlugins
 		void CFrequencySpectrumDisplayView::init()
 		{
 
-			OpenViBE::uint32 l_ui32ChannelCount = m_pBufferDatabase->m_pDimmensionSizes[0];
+			OpenViBE::uint32 l_ui32ChannelCount = static_cast<uint32>(m_pBufferDatabase->m_pDimmensionSizes[0]);
 			vector<string>& l_oChannelName = m_pBufferDatabase->m_pDimmesionLabels[0];
 
 			stringstream l_oLabelString;
